+++ tests/tests.ts
@@ -7,17 +7,17 @@ import { EOL } from 'os';
 describe('String.IsNullOrWhitespace', () => {
 
     it('should return true on null string', () => {
-        let teststring: string | null = null;
-        let result = String.IsNullOrWhiteSpace(teststring!);
+        const teststring: string | null = null;
+        let result = String.IsNullOrWhiteSpace(teststring);
         expect(result).to.equal(true);
-        result = String.isNullOrWhiteSpace(teststring!);
+        result = String.isNullOrWhiteSpace(teststring);
         expect(result).to.equal(true);
-        result = $String.isNullOrWhiteSpace(teststring!);
+        result = $String.isNullOrWhiteSpace(teststring);
         expect(result).to.equal(true);
     });
 
     it('should return true on empty string', () => {
-        let teststring = '';
+        const teststring = '';
         let result = String.IsNullOrWhiteSpace(teststring);
         expect(result).to.equal(true);
         result = String.isNullOrWhiteSpace(teststring);
@@ -27,7 +27,7 @@ describe('String.IsNullOrWhitespace', () => {
     });
 
     it('should return true only whitespace', () => {
-        let teststring = '    ';
+        const teststring = '    ';
         let result = String.IsNullOrWhiteSpace(teststring);
         expect(result).to.equal(true);
         result = String.isNullOrWhiteSpace(teststring);
@@ -37,7 +37,7 @@ describe('String.IsNullOrWhitespace', () => {
     });
 
     it('should return false contains non-whitespace characters', () => {
-        let teststring = '  s  ';
+        const teststring = '  s  ';
         let result = String.IsNullOrWhiteSpace(teststring);
         expect(result).to.equal(false);
         result = String.isNullOrWhiteSpace(teststring);
@@ -50,8 +50,8 @@ describe('String.IsNullOrWhitespace', () => {
 describe('String.Format Number Pattern', () => {
     describe('Placeholders', () => {
         it('should return template if no format is found', () => {
-            let template = "Bar";
-            let valueToInsert = "Foo";
+            const template = 'Bar';
+            const valueToInsert = 'Foo';
             let result = String.Format(template, valueToInsert);
             expect(result).to.equal(template);
             result = String.format(template, valueToInsert);
@@ -61,8 +61,8 @@ describe('String.Format Number Pattern', () => {
         });
 
         it('should format the string correct', () => {
-            let template = "{0}";
-            let valueToInsert = "Foo";
+            const template = '{0}';
+            const valueToInsert = 'Foo';
             let result = String.Format(template, valueToInsert);
             expect(result).to.equal(valueToInsert);
             result = String.format(template, valueToInsert);
@@ -72,9 +72,9 @@ describe('String.Format Number Pattern', () => {
         });
 
         it('should format the string correct multiple times', () => {
-            let template = "{0}Bar{0}";
-            let valueToInsert = "Foo";
-            let expectedValue = "FooBarFoo";
+            const template = '{0}Bar{0}';
+            const valueToInsert = 'Foo';
+            const expectedValue = 'FooBarFoo';
             let result = String.Format(template, valueToInsert);
             expect(result).to.equal(expectedValue);
             result = String.format(template, valueToInsert);
@@ -84,10 +84,10 @@ describe('String.Format Number Pattern', () => {
         });
 
         it('should format the string correct multiple values', () => {
-            let template = "{0}Bar{1}";
-            let valueToInsert = "Foo";
-            let secondValueToInsert = "Baz";
-            let expectedValue = "FooBarBaz";
+            const template = '{0}Bar{1}';
+            const valueToInsert = 'Foo';
+            const secondValueToInsert = 'Baz';
+            const expectedValue = 'FooBarBaz';
             let result = String.Format(template, valueToInsert, secondValueToInsert);
             expect(result).to.equal(expectedValue);
             result = String.format(template, valueToInsert, secondValueToInsert);
@@ -99,10 +99,10 @@ describe('String.Format Number Pattern', () => {
     describe('formating', () => {
         describe('dates', () => {
             it('should set the correct display date using Date', () => {
-                let template = "{0:d}";
-                let valueToInsert = new Date(2017, 4, 13);
+                const template = '{0:d}';
+                const valueToInsert = new Date(2017, 4, 13);
 
-                let expectedValue = "13.04.2017";
+                const expectedValue = '13.04.2017';
                 let result = String.Format(template, valueToInsert);
                 console.log(result);
                 expect(result).to.equal(expectedValue);
@@ -115,10 +115,10 @@ describe('String.Format Number Pattern', () => {
             });
 
             it('should set the correct sortable date using Date', () => {
-                let template = "{0:s}";
-                let valueToInsert = new Date(2017, 4, 13);
+                const template = '{0:s}';
+                const valueToInsert = new Date(2017, 4, 13);
 
-                let expectedValue = "2017-04-13";
+                const expectedValue = '2017-04-13';
                 let result = String.Format(template, valueToInsert);
                 console.log(result);
                 expect(result).to.equal(expectedValue);
@@ -131,10 +131,10 @@ describe('String.Format Number Pattern', () => {
             });
 
             it('should set the correct display date using string', () => {
-                let template = "{0:d}";
-                let valueToInsert = '2017-01-23 00:00';
+                const template = '{0:d}';
+                const valueToInsert = '2017-01-23 00:00';
 
-                let expectedValue = "23.01.2017";
+                const expectedValue = '23.01.2017';
                 let result = String.Format(template, valueToInsert);
                 console.log(result);
                 expect(result).to.equal(expectedValue);
@@ -147,10 +147,10 @@ describe('String.Format Number Pattern', () => {
             });
 
             it('should set the correct sortable date using string', () => {
-                let template = "{0:s}";
-                let valueToInsert = '21.03.2017 22:15:01';
+                const template = '{0:s}';
+                const valueToInsert = '21.03.2017 22:15:01';
 
-                let expectedValue = "2017-03-21T22:15:01";
+                const expectedValue = '2017-03-21T22:15:01';
                 let result = String.Format(template, valueToInsert);
                 console.log(result);
                 expect(result).to.equal(expectedValue);
@@ -163,10 +163,10 @@ describe('String.Format Number Pattern', () => {
             });
 
             it('should set the correct sortable date without time using string', () => {
-                let template = "{0:s}";
-                let valueToInsert = '21.03.2017';
+                const template = '{0:s}';
+                const valueToInsert = '21.03.2017';
 
-                let expectedValue = "2017-03-21T00:00:00";
+                const expectedValue = '2017-03-21T00:00:00';
                 let result = String.Format(template, valueToInsert);
                 console.log(result);
                 expect(result).to.equal(expectedValue);
@@ -181,9 +181,9 @@ describe('String.Format Number Pattern', () => {
 
         describe('uppercasing', () => {
             it('should return the string as uppercase', () => {
-                let expectedValue = 'AWESOME';
-                let template = '{0:U}';
-                let valueToInsert = 'awesome';
+                const expectedValue = 'AWESOME';
+                const template = '{0:U}';
+                const valueToInsert = 'awesome';
 
                 let actual = String.Format(template, valueToInsert);
                 expect(actual).to.equal(expectedValue);
@@ -194,9 +194,9 @@ describe('String.Format Number Pattern', () => {
             });
 
             it('should return the string as lowercase', () => {
-                let expectedValue = 'awesome';
-                let template = '{0:L}';
-                let valueToInsert = 'AWESOME';
+                const expectedValue = 'awesome';
+                const template = '{0:L}';
+                const valueToInsert = 'AWESOME';
 
                 let actual = String.Format(template, valueToInsert);
                 expect(actual).to.equal(expectedValue);
@@ -210,7 +210,7 @@ describe('String.Format Number Pattern', () => {
         describe('numbers', () => {
 
             it('should not pad without specifier using {0}', () => {
-                let template = '{0}';
+                const template = '{0}';
                 let result = String.Format(template, 5);
                 expect(result).to.equal('5');
                 result = String.format(template, 5);
@@ -220,7 +220,7 @@ describe('String.Format Number Pattern', () => {
             });
 
             it('should pad 5 to 05 using {0:00}', () => {
-                let template = '{0:00}';
+                const template = '{0:00}';
                 let result = String.Format(template, 5);
                 expect(result).to.equal('05');
                 result = String.format(template, 5);
@@ -230,7 +230,7 @@ describe('String.Format Number Pattern', () => {
             });
 
             it('should pad 5 to 005 using {0:000}', () => {
-                let template = '{0:000}';
+                const template = '{0:000}';
                 let result = String.Format(template, 5);
                 expect(result).to.equal('005');
                 result = String.format(template, 5);
@@ -240,7 +240,7 @@ describe('String.Format Number Pattern', () => {
             });
 
             it('should ignore padding when input is longer then template', () => {
-                let template = '{0:000}';
+                const template = '{0:000}';
                 let result = String.Format(template, 50000);
                 expect(result).to.equal('50000');
                 result = String.format(template, 50000);
@@ -250,9 +250,9 @@ describe('String.Format Number Pattern', () => {
             });
 
             it('should set the correct thousands seperator', () => {
-                let template = '{0:n}';
-                let valueToInsert = '10000000000';
-                let expectedValue = '10.000.000.000';
+                const template = '{0:n}';
+                const valueToInsert = '10000000000';
+                const expectedValue = '10.000.000.000';
 
                 let result = String.Format(template, valueToInsert);
                 expect(result).to.equal(expectedValue);
@@ -262,9 +262,9 @@ describe('String.Format Number Pattern', () => {
                 expect(result).to.equal(expectedValue);
             });
             it('should set the correct thousands seperator keeping the decimals', () => {
-                let template = '{0:n}';
-                let valueToInsert = '10000000000,12345';
-                let expectedValue = '10.000.000.000,12345';
+                const template = '{0:n}';
+                const valueToInsert = '10000000000,12345';
+                const expectedValue = '10.000.000.000,12345';
 
                 let result = String.Format(template, valueToInsert);
                 expect(result).to.equal(expectedValue);
@@ -335,77 +335,77 @@ describe('String.Format Number Pattern', () => {
 });
 
 describe('String.Format Text Pattern', () => {
-    describe("formatting", () => {
-        it("Should parse out the word", () => {
+    describe('formatting', () => {
+        it('Should parse out the word', () => {
             // Arrange
-            var fruit = { type: "apple", color: "red" };
+            const fruit = { type: 'apple', color: 'red' };
 
             // Act
             // Assert
-            let formatted = String.Format("the {type} is {color}", fruit);
-            expect(formatted).to.equal("the apple is red");
-            formatted = String.format("the {type} is {color}", fruit);
-            expect(formatted).to.equal("the apple is red");
-            formatted = $String.format("the {type} is {color}", fruit);
-            expect(formatted).to.equal("the apple is red");
+            let formatted = String.Format('the {type} is {color}', fruit);
+            expect(formatted).to.equal('the apple is red');
+            formatted = String.format('the {type} is {color}', fruit);
+            expect(formatted).to.equal('the apple is red');
+            formatted = $String.format('the {type} is {color}', fruit);
+            expect(formatted).to.equal('the apple is red');
         });
 
-        it("Should parse out the word with specifiers and TS Class", () => {
+        it('Should parse out the word with specifiers and TS Class', () => {
             // Arrange
-            var fruit: Fruit = new Fruit("apple", "RED", "31.12.2018 01:02:03", "10000");
+            const fruit: Fruit = new Fruit('apple', 'RED', '31.12.2018 01:02:03', '10000');
 
             // Act
             // Assert
-            let formatted = String.Format("the {type:U} is {color:L} shipped on {shippingDate:s} with an amount of {amount:n}", fruit);
-            expect(formatted).to.equal("the APPLE is red shipped on 2018-12-31T01:02:03 with an amount of 10.000");
-            formatted = String.format("the {type:U} is {color:L} shipped on {shippingDate:s} with an amount of {amount:n}", fruit);
-            expect(formatted).to.equal("the APPLE is red shipped on 2018-12-31T01:02:03 with an amount of 10.000");
-            formatted = $String.format("the {type:U} is {color:L} shipped on {shippingDate:s} with an amount of {amount:n}", fruit);
-            expect(formatted).to.equal("the APPLE is red shipped on 2018-12-31T01:02:03 with an amount of 10.000");
+            let formatted = String.Format('the {type:U} is {color:L} shipped on {shippingDate:s} with an amount of {amount:n}', fruit);
+            expect(formatted).to.equal('the APPLE is red shipped on 2018-12-31T01:02:03 with an amount of 10.000');
+            formatted = String.format('the {type:U} is {color:L} shipped on {shippingDate:s} with an amount of {amount:n}', fruit);
+            expect(formatted).to.equal('the APPLE is red shipped on 2018-12-31T01:02:03 with an amount of 10.000');
+            formatted = $String.format('the {type:U} is {color:L} shipped on {shippingDate:s} with an amount of {amount:n}', fruit);
+            expect(formatted).to.equal('the APPLE is red shipped on 2018-12-31T01:02:03 with an amount of 10.000');
         });
     });
 });
 
 describe('String.Join', () => {
     it('should join the given strings passed as args', () => {
-        let stringOne = "red", stringTwo = "yellow", stringThree = "blue";
+        const stringOne = 'red', stringTwo = 'yellow', stringThree = 'blue';
 
         let result = String.Join('; ', stringOne, stringTwo, stringThree);
-        expect(result).to.equal("red; yellow; blue");
+        expect(result).to.equal('red; yellow; blue');
         result = String.join('; ', stringOne, stringTwo, stringThree);
-        expect(result).to.equal("red; yellow; blue");
+        expect(result).to.equal('red; yellow; blue');
         result = $String.join('; ', stringOne, stringTwo, stringThree);
-        expect(result).to.equal("red; yellow; blue");
+        expect(result).to.equal('red; yellow; blue');
     });
 
     it('should join the given array', () => {
-        let object = ["red", "yellow", "blue"];
+        const object = ['red', 'yellow', 'blue'];
         let result = String.Join('; ', object);
-        expect(result).to.equal("red; yellow; blue");
+        expect(result).to.equal('red; yellow; blue');
         result = String.join('; ', object);
-        expect(result).to.equal("red; yellow; blue");
+        expect(result).to.equal('red; yellow; blue');
         result = $String.join('; ', object);
-        expect(result).to.equal("red; yellow; blue");
+        expect(result).to.equal('red; yellow; blue');
     });
 
     it('should join the given object', () => {
-        let object = { Name: "Foo", Value: "Bar" };
+        const object = { Name: 'Foo', Value: 'Bar' };
 
         let result = String.Join('.', object);
         console.log(result);
-        expect(result).to.equal("Foo.Bar");
+        expect(result).to.equal('Foo.Bar');
         result = String.join('.', object);
         console.log(result);
-        expect(result).to.equal("Foo.Bar");
+        expect(result).to.equal('Foo.Bar');
         result = $String.join('.', object);
         console.log(result);
-        expect(result).to.equal("Foo.Bar");
+        expect(result).to.equal('Foo.Bar');
     });
 });
 
 describe('StringBuilder initialization', () => {
     it('should not add empty string if there is no ctor parameter', () => {
-        var builder = new StringBuilder();
+        const builder = new StringBuilder();
         builder.Append('First Part... ');
         builder.Append('Second Part...');
 
@@ -414,7 +414,7 @@ describe('StringBuilder initialization', () => {
     });
 
     it('should add a string if there is ctor parameter', () => {
-        var builder = new StringBuilder($String.format("First {0}... ", "Part"));
+        const builder = new StringBuilder($String.format('First {0}... ', 'Part'));
         builder.AppendFormat('Second {0}...', 'Part');
 
         console.log(builder.ToString());
@@ -425,7 +425,7 @@ describe('StringBuilder initialization', () => {
 
 describe('StringBuilderng.Append', () => {
     it('should append characters', () => {
-        var builder = new StringBuilder();
+        const builder = new StringBuilder();
         builder.Append('First Part... ');
         builder.Append('Second Part...');
 
@@ -434,7 +434,7 @@ describe('StringBuilderng.Append', () => {
     });
 
     it('should append characters', () => {
-        var builder = new StringBuilder();
+        const builder = new StringBuilder();
         builder.AppendFormat('First {0}... ', 'Part');
         builder.AppendFormat('Second {0}...', 'Part');
 
@@ -446,7 +446,7 @@ describe('StringBuilderng.Append', () => {
 
 describe('StringBuilder.AppendLine', () => {
     it('should append characters and new line', () => {
-        var builder = new StringBuilder();
+        const builder = new StringBuilder();
         builder.AppendLine('First Line...');
         builder.AppendLine('Second Line...');
 
@@ -455,7 +455,7 @@ describe('StringBuilder.AppendLine', () => {
     });
 
     it('should append characters and new line', () => {
-        var builder = new StringBuilder();
+        const builder = new StringBuilder();
         builder.AppendLineFormat('First {0}...', 'Line');
         builder.AppendLineFormat('Second {0}...', 'Line');
 
@@ -465,7 +465,7 @@ describe('StringBuilder.AppendLine', () => {
     });
 
     it('should append characters and new line', () => {
-        var builder = new StringBuilder();
+        const builder = new StringBuilder();
         builder.AppendLine('First Line...');
         builder.AppendLine('Second Line...');